Add an implementation using Zapier CLI.
This is an alternative to UI-based configuration. See
https://platform.zapier.com/quickstart/cli-tutorial for using it. This
code was initialized from the UI-based verison using "zapier convert".
Then it was modified to support self-hosted domains.

Tests boilerplate was generated, with only the bare minimum added to
get test to pass with "zapier test". Running tests that pass requires
running Grist locally at localhost:8080, and passing TEST_GRIST_API_KEY
and TEST_GRIST_DOC_ID in the environment, with the docID of a local
document that's a copy of
https://templates.getgrist.com/doc/lightweight-crm template.